requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.3.0
async-timeout==4.0.3
backcall==0.2.0
cachetools==5.3.1
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
decorator==5.1.1
Django==4.2.4
django-redis==5.3.0
djangorestframework==3.14.0
executing==1.2.0
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-texttospeech==2.14.1
google-cloud-translate==3.12.0
googleapis-common-protos==1.60.0
googletrans==4.0.0rc1
grpcio==1.57.0
grpcio-status==1.57.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ipython==8.15.0
jedi==0.19.0
matplotlib-inline==0.1.6
numpy==1.25.2
pandas==2.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3
redis==5.0.0
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
